+++ mocket/core/socket.py
@@ -10,8 +10,8 @@ from typing import Any, Type
 
 from typing_extensions import Self
 
+from mocket.core.io import MocketSocketIO
 from mocket.entry import MocketEntry
-from mocket.io import MocketSocketIO
 from mocket.mocket import Mocket
 from mocket.mode import MocketMode
 from mocket.types import (
